Split character-workspace out of its router (923 -> 59 lines)
character-interact had both the chat and adventure modes inlined, so the
component whose only job is choosing between browsing and working was 923
lines -- 482 of template, 440 of script. dream-interact does the same job
in 57. All of it moved to character-workspace.vue unchanged; nothing about
how a Character conversation or adventure works changed here.

  <x-gallery v-if="!selected" />     <x-workspace v-else />

Per Silas, the interact tier "is where we are actually hitting what we do
with them uniquely", so the shared thing is the frame, not the contents.

WHY THE SESSION SIGNAL COMES FROM THE STORE. Interact mode is not simply
`Boolean(selectedCharacter)` -- a conversation already under way keeps you
in the workspace after the Character is deselected. That condition was
answered by a local `chatMessages` array, which is display state and had to
move out with the workspace, leaving the router unable to see its own
routing condition. The workspace now registers each real Chat id with
chatStore under the 'character' scope, the same answer reward-interact and
bot-interact use.

The narrative-kit counter held at 7/7 across this split without being
touched, which is the earlier graph-walk fix doing its job: the hand-kept
INTERMEDIARIES list it replaced would have reported this refactor as a
regression, because the chat window and choice list moved one hop down.

Rule 5 ratchets 923 -> 59. Note for the next step: `--update` writes raw
JSON.stringify, so the baseline needs `prettier --write` after it or the
diff is 59 lines of re-wrapping instead of the one line that changed.
